test_projects/chat_wifi/host/requirements.txt
# Web server
websockets>=12.0
aiohttp>=3.9.0

# Audio processing
numpy>=1.24.0
soundfile>=0.12.0

# Speech-to-Text (choose one)
openai-whisper>=20231117
# faster-whisper>=0.10.0  # Alternative, faster

# Grok/xAI API
openai>=1.0.0  # Compatible client for Grok API

# Text-to-Speech
# piper-tts>=1.2.0  # Local TTS (install separately)

# Utilities
python-dotenv>=1.0.0
